ConfigMind /requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
boto3>=1.34.0
pyyaml>=6.0.1
pydantic>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
